// src/api/model/bankResponseDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * accounting
 * Accounting API description
 * OpenAPI spec version: 1.0
 */

export interface BankResponseDto {
  id: string;
  name: string;
  number?: string;
  iban?: string;
  swift?: string;
  currency: string;
  default: boolean;
}